sentiment-analysis-process/requirements.txt
transformers==4.27.3
tokenizers==0.13.2
tqdm==4.65.0
tensorflow==2.11.0
protobuf==3.19.6
pandas==1.5.3
numpy==1.22.4
grpcio-tools==1.30.0
grpcio==1.53.0